backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
